Add test case for bad ManyToMany declaration (through entity not defined)
