+++ tests/Schema/Driver/Postgres/RenderRelationsTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\Schema\Tests\Driver\Postgres;
+
+use Cycle\Schema\Tests\Generator\RenderRelationsTest as BaseTest;
+
+class RenderRelationsTest extends BaseTest
+{
+    public const DRIVER = 'postgres';
+}
